+++ NEWS.md
@@ -1,13 +1,21 @@
 
+# archeofrag 1.2.4
+
+* Add `frag.cohesion.ranking()` function.
+
+
+# archeofrag 1.2.3
+Released: 2025-09-26
+
+* Slight change to anticipate an upcoming change in the `igraph` package.
+
 # archeofrag 1.2.2
-Released: 
 
 * Add 'St Cesaire 1987' and 'St Cesaire 2021' datasets.
 * Add the 'Grotte 16' dataset.
 * Add the 'Geelbek' dataset.
 
 # archeofrag 1.2.1
-Released: 
 
 * Add 'Grande Rivoire 1st Meso' and 'Grande Rivoire 2nd Meso' datasets. Remove the 'Grande Rivoire' dataset.
 * Code optimisation: add internal sub-functions `.fragments.balance()` and `.components.balance()`. The first is now used both in `frag.get.parameters()` and `frag.graph.reduce()` (in the last case, to avoid calling `frag.get.parameters()` in `frag.graph.reduce()`).